% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_growth_interval.R, R/s3_curve_test.R,
%   R/s3_methods.R, R/tidy_reexport.R
\name{tidy.phytogrow_classical}
\alias{tidy.phytogrow_classical}
\alias{tidy.phytogrow_curve_test}
\alias{tidy.phytogrow_fit}
\alias{tidy}
\title{Tidy Object Results}
\usage{
\method{tidy}{phytogrow_classical}(x, ci_source = c("auto", "classical", "bootstrap", "all"), ...)

\method{tidy}{phytogrow_curve_test}(x, ...)

\method{tidy}{phytogrow_fit}(x, ...)

tidy(x, ...)
}
\arguments{
\item{x}{Object to tidy.}

\item{ci_source}{For \code{phytogrow_classical}, choose \code{"auto"}, \code{"classical"},
\code{"bootstrap"}, or \code{"all"} to filter returned rows.}

\item{...}{Additional arguments passed to method implementations.}
}
\value{
A tibble-like object depending on the method.
}
\description{
Re-export of \code{\link[generics:tidy]{generics::tidy()}} for objects from \code{phytogrowR}.
}
