% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{phytogrowR-package}
\alias{phytogrowR}
\alias{phytogrowR-package}
\title{phytogrowR: Modern Plant Growth Analysis}
\description{
\code{phytogrowR} supports classical and functional growth analysis for plant
physiology and agronomic experiments based on repeated measurements,
destructive sequential harvests, biomass trajectories, and leaf area series.
Core concepts are based on classical growth analysis references including
Hunt (1990) and Benincasa (2003).
}
\references{
Hunt, R. (1990). \emph{Basic Growth Analysis}. Unwin Hyman, London.

Benincasa, M. M. P. (2003). \emph{Plant Growth Analysis: Basic Concepts}.
FUNEP, Jaboticabal.

Hunt, R., Causton, D. R., Shipley, B., and Askew, A. P. (2002).
A modern tool for classical plant growth analysis.
\emph{Annals of Botany}, 90(4), 485-488. doi:10.1093/aob/mcf214
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agrobioestat/phytogrowR}
  \item Report bugs at \url{https://github.com/agrobioestat/phytogrowR/issues}
}

}
\author{
\strong{Maintainer}: Joao Everthon da Silva Ribeiro \email{everthonribeiro10@gmail.com}

}
\keyword{internal}
